Create ERC721.sol
This code defines a contract called ERC721 that implements the ERC721 standard and the cactus_core_api_erc721 interface provided by the Hyperledger Cactus framework. In the constructor, the name and symbol of the token are set.

To create a new ERC721 token, you can deploy this contract on the Ethereum blockchain using a tool such as Truffle. Once deployed, you can use the functions provided by the ERC721 and cactus_core_api_erc721 interfaces to manage your token. For example, you can use the mint() function to create new tokens and the transferFrom() function to transfer tokens to other addresses.