Create SecurityToken.sol
This contract implements the ERC1400 standard for security tokens, which defines a set of functions and events that allow for the issuance, transfer, and redemption of security tokens. The SecurityToken contract has a name, symbol, and decimals field to define the token's metadata, as well as a totalSupply field to track the total number of tokens in circulation. It also has an owner field to track the address of the token owner